+++ src/main/java/com/techfork/personalization/application/PersonalizationProfileAnalysis.java
@@ -1,0 +1,9 @@
+package com.techfork.personalization.application;
+
+import java.util.List;
+
+public record PersonalizationProfileAnalysis(
+        String profileText,
+        List<String> keyKeywords
+) {
+}
