+++ src/test/java/com/techfork/personalization/application/PersonalizationProfileServiceTest.java
@@ -2,7 +2,6 @@ package com.techfork.personalization.application;
 
 import com.techfork.domain.recommendation.service.RecommendationService;
 import com.techfork.global.llm.EmbeddingClient;
-import com.techfork.global.llm.LlmClient;
 import com.techfork.personalization.infrastructure.PersonalizationProfileDocument;
 import com.techfork.personalization.infrastructure.PersonalizationProfileDocumentRepository;
 import com.techfork.useraccount.domain.User;
@@ -34,6 +33,9 @@ class PersonalizationProfileServiceTest {
     private UserActivityCollector userActivityCollector;
 
     @Mock
+    private PersonalizationProfileAnalyzer personalizationProfileAnalyzer;
+
+    @Mock
     private PersonalizationProfileDocumentRepository personalizationProfileDocumentRepository;
 
     @Mock
@@ -43,43 +45,30 @@ class PersonalizationProfileServiceTest {
     private RecommendationService recommendationService;
 
     @Mock
-    private LlmClient llmClient;
-
-    @Mock
     private EmbeddingClient embeddingClient;
 
     @InjectMocks
     private PersonalizationProfileService personalizationProfileService;
 
     @Test
-    @DisplayName("사용자 활동 데이터를 모아 개인화 프로필을 생성하고 저장한다")
-    void generatePersonalizationProfileSync_CollectsActivityDataParsesAndSavesProfile() {
+    @DisplayName("사용자 활동 데이터를 분석해 개인화 프로필을 저장하고 추천을 생성한다")
+    void generatePersonalizationProfileSync_AnalyzesActivityDataSavesProfileAndGeneratesRecommendations() {
         Long userId = 1L;
         User user = createUser(userId);
         UserActivityData activityData = activityData(
                 List.of("Java", "Spring", "Docker"),
-                List.of(
-                        postActivityData("30초 포스트", List.of("Java"), "가볍게 훑어봄"),
-                        postActivityData("90초 포스트", List.of("Spring"), "빠르게 읽음"),
-                        postActivityData("300초 포스트", List.of("JPA"), "읽음"),
-                        postActivityData("600초 포스트", List.of("Kafka"), "정독함"),
-                        postActivityData("601초 포스트", List.of("Docker"), "깊게 읽음"),
-                        postActivityData("null 포스트", List.of("Elastic"), "읽음")
-                ),
+                List.of(postActivityData("읽은 포스트", List.of("Java"), "정독함")),
                 List.of(postActivityData("북마크 포스트", List.of("Kubernetes", "Helm"), null)),
                 List.of("Spring Batch", "Elasticsearch vector")
         );
+        PersonalizationProfileAnalysis analysis = profileAnalysis(
+                "Java와 Spring 기반 백엔드, Docker 중심 운영 자동화, Elasticsearch 검색 최적화에 집중하는 사용자",
+                List.of("Java", "Spring", "Docker", "Elasticsearch", "Batch")
+        );
 
         given(userActivityCollector.collect(userId)).willReturn(activityData);
-        given(llmClient.call(anyString(), anyString()))
-                .willReturn("""
-                        ### PROFILE
-                        Java와 Spring 기반 백엔드, Docker 중심 운영 자동화, Elasticsearch 검색 최적화에 집중하는 사용자
-
-                        ### KEYWORDS
-                        Java, Spring, Docker, Elasticsearch, Batch
-                        """);
-        given(embeddingClient.embed(anyString())).willReturn(List.of(0.1f, 0.2f, 0.3f));
+        given(personalizationProfileAnalyzer.analyze(activityData)).willReturn(analysis);
+        given(embeddingClient.embed(analysis.profileText())).willReturn(List.of(0.1f, 0.2f, 0.3f));
         given(personalizationProfileDocumentRepository.save(any(PersonalizationProfileDocument.class)))
                 .willAnswer(invocation -> invocation.getArgument(0));
         given(userRepository.findById(userId)).willReturn(Optional.of(user));
@@ -87,100 +76,21 @@ class PersonalizationProfileServiceTest {
 
         personalizationProfileService.generatePersonalizationProfileSync(userId);
 
-        ArgumentCaptor<String> promptCaptor = ArgumentCaptor.forClass(String.class);
-        verify(llmClient).call(anyString(), promptCaptor.capture());
-
-        String prompt = promptCaptor.getValue();
-        assertThat(prompt)
-                .contains("Java")
-                .contains("Spring")
-                .contains("Docker")
-                .contains("30초 포스트")
-                .contains("90초 포스트")
-                .contains("300초 포스트")
-                .contains("600초 포스트")
-                .contains("601초 포스트")
-                .contains("null 포스트")
-                .contains("북마크 포스트")
-                .contains("Spring Batch")
-                .contains("Elasticsearch vector")
-                .contains("가볍게 훑어봄")
-                .contains("빠르게 읽음")
-                .contains("읽음")
-                .contains("정독함")
-                .contains("깊게 읽음");
-
         ArgumentCaptor<PersonalizationProfileDocument> documentCaptor = ArgumentCaptor.forClass(PersonalizationProfileDocument.class);
         verify(personalizationProfileDocumentRepository).save(documentCaptor.capture());
 
         PersonalizationProfileDocument savedDocument = documentCaptor.getValue();
         assertThat(savedDocument.getUserId()).isEqualTo(userId);
-        assertThat(savedDocument.getProfileText())
-                .isEqualTo("Java와 Spring 기반 백엔드, Docker 중심 운영 자동화, Elasticsearch 검색 최적화에 집중하는 사용자");
+        assertThat(savedDocument.getProfileText()).isEqualTo(analysis.profileText());
         assertThat(savedDocument.getProfileVector()).containsExactly(0.1f, 0.2f, 0.3f);
         assertThat(savedDocument.getInterests()).containsExactly("Java", "Spring", "Docker");
-        assertThat(savedDocument.getKeyKeywords())
-                .containsExactly("Java", "Spring", "Docker", "Elasticsearch", "Batch");
+        assertThat(savedDocument.getKeyKeywords()).containsExactlyElementsOf(analysis.keyKeywords());
 
         verify(userActivityCollector).collect(userId);
+        verify(personalizationProfileAnalyzer).analyze(activityData);
+        verify(embeddingClient).embed(analysis.profileText());
         verify(userRepository).findById(userId);
         verify(recommendationService).generateRecommendationsForUser(user);
-    }
-
-    @Test
-    @DisplayName("LLM 응답을 파싱하지 못하면 전체 텍스트를 프로필로 fallback 저장한다")
-    void generatePersonalizationProfileSync_FallsBackToFullTextWhenSectionsAreMissing() {
-        Long userId = 2L;
-        User user = createUser(userId);
-        String llmResponse = "섹션 없이도 전체 응답을 개인화 프로필로 저장해야 한다";
-
-        given(userActivityCollector.collect(userId)).willReturn(emptyActivityData());
-        given(llmClient.call(anyString(), anyString())).willReturn(llmResponse);
-        given(embeddingClient.embed(llmResponse)).willReturn(List.of(1.0f, 2.0f));
-        given(personalizationProfileDocumentRepository.save(any(PersonalizationProfileDocument.class)))
-                .willAnswer(invocation -> invocation.getArgument(0));
-        given(userRepository.findById(userId)).willReturn(Optional.of(user));
-
-        personalizationProfileService.generatePersonalizationProfileSync(userId);
-
-        ArgumentCaptor<PersonalizationProfileDocument> documentCaptor = ArgumentCaptor.forClass(PersonalizationProfileDocument.class);
-        verify(personalizationProfileDocumentRepository).save(documentCaptor.capture());
-
-        PersonalizationProfileDocument savedDocument = documentCaptor.getValue();
-        assertThat(savedDocument.getProfileText()).isEqualTo(llmResponse);
-        assertThat(savedDocument.getKeyKeywords()).isEmpty();
-        assertThat(savedDocument.getProfileVector()).containsExactly(1.0f, 2.0f);
-        verify(userActivityCollector).collect(userId);
-    }
-
-    @Test
-    @DisplayName("LLM 응답의 핵심 키워드는 최대 5개까지만 저장한다")
-    void generatePersonalizationProfileSync_LimitsKeyKeywordsToFive() {
-        Long userId = 4L;
-        User user = createUser(userId);
-
-        given(userActivityCollector.collect(userId)).willReturn(emptyActivityData());
-        given(llmClient.call(anyString(), anyString()))
-                .willReturn("""
-                        ### PROFILE
-                        Java와 Spring 기반 백엔드 성능 최적화에 집중하는 사용자
-
-                        ### KEYWORDS
-                        Java, Spring, JPA, Redis, Kafka, Kubernetes, Elasticsearch
-                        """);
-        given(embeddingClient.embed(anyString())).willReturn(List.of(0.4f, 0.5f));
-        given(personalizationProfileDocumentRepository.save(any(PersonalizationProfileDocument.class)))
-                .willAnswer(invocation -> invocation.getArgument(0));
-        given(userRepository.findById(userId)).willReturn(Optional.of(user));
-
-        personalizationProfileService.generatePersonalizationProfileSync(userId);
-
-        ArgumentCaptor<PersonalizationProfileDocument> documentCaptor = ArgumentCaptor.forClass(PersonalizationProfileDocument.class);
-        verify(personalizationProfileDocumentRepository).save(documentCaptor.capture());
-
-        assertThat(documentCaptor.getValue().getKeyKeywords())
-                .containsExactly("Java", "Spring", "JPA", "Redis", "Kafka");
-        verify(userActivityCollector).collect(userId);
     }
 
     @Test
@@ -188,16 +98,14 @@ class PersonalizationProfileServiceTest {
     void generatePersonalizationProfileSync_RecommendationFailureDoesNotBreakProfileSave() {
         Long userId = 3L;
         User user = createUser(userId);
+        UserActivityData activityData = emptyActivityData();
+        PersonalizationProfileAnalysis analysis = profileAnalysis(
+                "추천 실패와 무관하게 저장되어야 하는 프로필",
+                List.of("테스트", "회귀")
+        );
 
-        given(userActivityCollector.collect(userId)).willReturn(emptyActivityData());
-        given(llmClient.call(anyString(), anyString()))
-                .willReturn("""
-                        ### PROFILE
-                        추천 실패와 무관하게 저장되어야 하는 프로필
-
-                        ### KEYWORDS
-                        테스트, 회귀
-                        """);
+        given(userActivityCollector.collect(userId)).willReturn(activityData);
+        given(personalizationProfileAnalyzer.analyze(activityData)).willReturn(analysis);
         given(embeddingClient.embed(anyString())).willReturn(List.of(9.0f, 8.0f));
         given(personalizationProfileDocumentRepository.save(any(PersonalizationProfileDocument.class)))
                 .willAnswer(invocation -> invocation.getArgument(0));
@@ -210,6 +118,7 @@ class PersonalizationProfileServiceTest {
 
         verify(personalizationProfileDocumentRepository).save(any(PersonalizationProfileDocument.class));
         verify(userActivityCollector).collect(userId);
+        verify(personalizationProfileAnalyzer).analyze(activityData);
         verify(recommendationService).generateRecommendationsForUser(user);
     }
 
@@ -228,6 +137,10 @@ class PersonalizationProfileServiceTest {
 
     private PostActivityData postActivityData(String title, List<String> keywords, String readingEngagement) {
         return new PostActivityData(title, keywords, readingEngagement);
+    }
+
+    private PersonalizationProfileAnalysis profileAnalysis(String profileText, List<String> keyKeywords) {
+        return new PersonalizationProfileAnalysis(profileText, keyKeywords);
     }
 
     private User createUser(Long userId) {
